GPU: Fix rare race condition
[Detail]
This patch moves the call to kbase_pm_context_idle() in kbase_destroy_context() to the end of the function.
This ensures that the GPU is powered on for the entire function, removing a rare race condition between the
GPU being powered off and the call to kbase_ctx_sched_remove_ctx(), which could cause a kernel panic.

MTK-Commit-Id: bd552d88832d3f259dfa80a19850d2614d629a1c

Change-Id: Ia109463c02aefeeaf19ee8fcb1659313f75cd2da
Signed-off-by: rita.chan <rita.chan@mediatek.com>
CR-Id: ALPS03446607
Feature: Power Management
Signed-off-by: dodyirawan85 <40514988+dodyirawan85@users.noreply.github.com>